worker/requirements.txt
supabase
python-dotenv
transformers
torch
sentencepiece